+++ src/pages/rooms-overview/index.vue
@@ -42,7 +42,12 @@
 				</div>
 			</v-col>
 		</v-row>
-		<vCustomDialog v-model="groupDialog.isOpen" class="custom-dialog">
+		<vCustomDialog
+			ref="custom-dialog"
+			v-model="groupDialog.isOpen"
+			class="custom-dialog"
+			@dialog-closed="groupDialog.groupData = {}"
+		>
 			<div slot="title">
 				<v-text-field
 					v-show="roomNameEditMode"
@@ -67,17 +72,20 @@
 			<template slot="content">
 				<v-row class="d-flex justify-center ma-1">
 					<v-col
-						v-for="item in groupDialog.groupData.groupElements"
+						v-for="(item, index) in groupDialog.groupData.groupElements"
 						:key="item.id"
 						class="d-flex justify-center"
 						:cols="maxItem"
 					>
 						<vRoomAvatar
+							:ref="`index-${index}`"
 							:item="item"
-							:size="(dimensions.cellWidth * ratios.itemRatio) / 2"
+							:size="dimensions.cellWidth * ratios.itemRatio"
+							:is-group-avatar-list="true"
 							:show-badge="true"
-							class="rounded-xl"
+							class="rounded-xl dialog-avatar"
 							show-sub-title
+							@startDrag="dragFromGroup"
 						></vRoomAvatar>
 					</v-col>
 				</v-row>
@@ -132,6 +140,7 @@ export default {
 			mdiPencil,
 			mdiKeyboardReturn,
 			roomNameEditMode: false,
+			draggedElementName: "",
 		};
 	},
 	computed: {
@@ -146,7 +155,7 @@ export default {
 		},
 	},
 	async created() {
-		await RoomsModule.fetch(); // this method will receive a string parameter (Eg, mobile | tablet | desktop)
+		await RoomsModule.fetch(); // TODO: this method will receive a string parameter (Eg, mobile | tablet | desktop)
 		this.roomsData = RoomsModule.getRoomsData;
 		this.getDeviceDims();
 	},
@@ -195,56 +204,62 @@ export default {
 			this.draggedElement.to = null;
 			this.draggedElement.item = element;
 			this.showDeleteSection = true;
+			this.draggedElementName = this.getElementNameByRef(pos);
 		},
-		async setDropElement(pos) {
+		setDropElement(pos) {
 			this.draggedElement.to = pos;
-			const fromElementName = this.getElementNameByRef(
-				this.draggedElement.from
-			);
 			const toElementName = this.getElementNameByRef(pos);
 
 			if (
-				(fromElementName == "vRoomAvatar" || "vRoomGroupAvatar") &&
+				(this.draggedElementName == "vRoomAvatar" ||
+					"vRoomGroupAvatar" ||
+					"groupItem") &&
 				toElementName == "vRoomEmptyAvatar"
 			) {
-				await RoomsModule.align(this.draggedElement);
-				this.roomsData = RoomsModule.getRoomsData;
+				this.savePosition();
 			}
 			this.showDeleteSection = false;
 		},
-		async setGroupElements(pos) {
+		setGroupElements(pos) {
 			this.draggedElement.to = pos;
-			const fromElementName = this.getElementNameByRef(
-				this.draggedElement.from
-			);
-			const toElementName = this.getElementNameByRef(pos);
-
-			if (this.draggedElement.from == pos) return;
-
-			if (fromElementName == "vRoomAvatar" && toElementName == "vRoomAvatar") {
-				await RoomsModule.align(this.draggedElement);
-				this.roomsData = RoomsModule.getRoomsData;
-			}
-		},
-		async addGroupElements(pos) {
-			this.draggedElement.to = pos;
-			const fromElementName = this.getElementNameByRef(
-				this.draggedElement.from
-			);
 			const toElementName = this.getElementNameByRef(pos);
 
 			if (this.draggedElement.from == pos) return;
 
 			if (
-				fromElementName == "vRoomAvatar" &&
+				(this.draggedElementName == "vRoomAvatar" || "groupItem") &&
+				toElementName == "vRoomAvatar"
+			) {
+				this.savePosition();
+			}
+		},
+		addGroupElements(pos) {
+			this.draggedElement.to = pos;
+			const toElementName = this.getElementNameByRef(pos);
+
+			if (this.draggedElement.from == pos) return;
+
+			if (
+				(this.draggedElementName == "vRoomAvatar" || "groupItem") &&
 				toElementName == "vRoomGroupAvatar"
 			) {
-				await RoomsModule.align(this.draggedElement);
-				this.roomsData = RoomsModule.getRoomsData;
+				this.savePosition();
 			}
 		},
 		getElementNameByRef(pos) {
 			return this.$refs[`${pos.y}-${pos.x}`][0].$options["_componentTag"];
+		},
+		dragFromGroup(element) {
+			this.draggedElement.from = {
+				x: this.groupDialog.groupData.xPosition,
+				y: this.groupDialog.groupData.yPosition,
+				groupIndex: this.roomsData
+					.find((item) => item.id == this.groupDialog.groupData.id)
+					.groupElements.findIndex((groupItem) => groupItem.id == element.id),
+			};
+			this.draggedElement.item = element;
+			this.draggedElementName = "groupItem";
+			this.groupDialog.isOpen = false;
 		},
 		deleteAvatar() {
 			// TODO: delete event will be here
@@ -270,6 +285,16 @@ export default {
 		*/
 		onRoomNameInputEnter(event) {
 			event.target.blur();
+		},
+		setGroupElementIndex(elementId) {
+			return this.roomsData
+				.find((item) => item.id == this.groupDialog.groupData.id)
+				.groupElements.findIndex((groupItem) => groupItem.id == elementId);
+		},
+		async savePosition() {
+			await RoomsModule.align(this.draggedElement);
+			this.roomsData = RoomsModule.getRoomsData;
+			this.groupDialog.groupData = {};
 		},
 	},
 };